@@ -18,92 +18,92 @@
   },
   "dependencies": {
     "@ant-design/icons": "^4.7.0",
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.3.0",
-    "@etherdata-blockchain/common": "^1.13.0",
-    "@etherdata-blockchain/history": "^1.13.0",
-    "@etherdata-blockchain/logger": "^1.13.0",
-    "@etherdata-blockchain/next-js-handlers": "^1.13.0",
-    "@etherdata-blockchain/services": "^1.13.0",
-    "@etherdata-blockchain/storage-model": "^1.13.0",
-    "@mui/icons-material": "^5.0.4",
-    "@mui/lab": "^5.0.0-alpha.55",
-    "@mui/material": "^5.0.4",
-    "@mui/styles": "^5.0.1",
-    "@mui/x-data-grid": "^5.0.1",
-    "@rjsf/antd": "^3.2.1",
-    "@rjsf/bootstrap-4": "^3.2.1",
-    "@rjsf/core": "^3.2.1",
+    "@emotion/react": "^11.8.1",
+    "@emotion/styled": "^11.8.1",
+    "@etherdata-blockchain/common": "^1.15.0",
+    "@etherdata-blockchain/history": "^1.15.0",
+    "@etherdata-blockchain/logger": "^1.15.0",
+    "@etherdata-blockchain/next-js-handlers": "^1.15.0",
+    "@etherdata-blockchain/services": "^1.15.0",
+    "@etherdata-blockchain/storage-model": "^1.15.0",
+    "@mui/icons-material": "^5.4.2",
+    "@mui/lab": "^5.0.0-alpha.70",
+    "@mui/material": "^5.4.3",
+    "@mui/styles": "^5.4.2",
+    "@mui/x-data-grid": "^5.5.1",
+    "@rjsf/antd": "^4.0.0",
+    "@rjsf/bootstrap-4": "^4.0.0",
+    "@rjsf/core": "^4.0.0",
     "@types/adm-zip": "^0.4.34",
     "@types/css-mediaquery": "^0.1.1",
     "@types/express": "^4.17.13",
-    "@types/file-saver": "^2.0.4",
-    "@types/jsonwebtoken": "^8.5.4",
+    "@types/file-saver": "^2.0.5",
+    "@types/jsonwebtoken": "^8.5.8",
     "@types/lodash": "^4.14.178",
     "@types/randomcolor": "^0.5.6",
     "@types/url-join": "^4.0.1",
-    "@types/uuid": "^8.3.1",
+    "@types/uuid": "^8.3.4",
     "adm-zip": "^0.5.9",
-    "antd": "^4.17.2",
+    "antd": "^4.18.8",
     "axios": "^0.26.0",
-    "bootstrap": "5.1.0",
+    "bootstrap": "5.1.3",
     "css-mediaquery": "^0.1.2",
     "dayjs": "^1.10.7",
     "etd-react-ui": "^2.0.1",
-    "express": "^4.17.1",
+    "express": "^4.17.3",
     "file-saver": "^2.0.5",
     "formik": "^2.2.9",
     "http-method-enum": "^1.0.0",
-    "http-status-codes": "^2.1.4",
+    "http-status-codes": "^2.2.0",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.21",
     "material-ui-popup-state": "^2.0.0",
     "moment": "^2.29.1",
-    "next": "12.0.10",
-    "nextjs-progressbar": "^0.0.11",
-    "query-string": "^7.0.1",
+    "next": "12.1.0",
+    "nextjs-progressbar": "^0.0.13",
+    "query-string": "^7.1.1",
     "randomcolor": "^0.6.2",
     "react": "17.0.2",
-    "react-bootstrap": "^2.0.2",
+    "react-bootstrap": "^2.1.2",
     "react-dom": "17.0.2",
-    "realm-web": "^1.3.0",
-    "recharts": "^2.0.9",
+    "realm-web": "^1.7.0",
+    "recharts": "^2.1.9",
     "socket.io": "^4.4.1",
     "socket.io-client": "^4.4.1",
     "start-server-and-test": "^1.14.0",
-    "styled-components": "^5.3.0",
-    "swr": "^1.1.1",
-    "ts-jest": "^27.0.7",
-    "ts-node": "^10.1.0",
-    "typescript-json-schema": "^0.52.0",
+    "styled-components": "^5.3.3",
+    "swr": "^1.2.2",
+    "ts-jest": "^27.1.3",
+    "ts-node": "^10.5.0",
+    "typescript-json-schema": "^0.53.0",
     "url-join": "^4.0.1"
   },
   "devDependencies": {
     "@babel/plugin-syntax-jsx": "^7.16.7",
-    "@babel/preset-env": "^7.16.4",
-    "@babel/preset-react": "^7.16.0",
-    "@babel/preset-typescript": "^7.14.5",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
     "@release-it/conventional-changelog": "^4.1.0",
-    "@shelf/jest-mongodb": "^2.0.1",
-    "@testing-library/dom": "^8.1.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@types/jest": "^27.0.2",
-    "@types/react": "17.0.14",
-    "@typescript-eslint/eslint-plugin": "^5.4.0",
-    "@typescript-eslint/parser": "^5.4.0",
-    "babel-jest": "^27.3.1",
-    "eslint": "^7.32.0",
+    "@shelf/jest-mongodb": "^2.2.0",
+    "@testing-library/dom": "^8.11.3",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.3",
+    "@types/jest": "^27.4.0",
+    "@types/react": "17.0.39",
+    "@typescript-eslint/eslint-plugin": "^5.12.1",
+    "@typescript-eslint/parser": "^5.12.1",
+    "babel-jest": "^27.5.1",
+    "eslint": "^8.9.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-config-next": "12.0.4",
-    "eslint-plugin-react": "^7.27.1",
+    "eslint-config-next": "12.1.0",
+    "eslint-plugin-react": "^7.28.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.0.6",
-    "node-mocks-http": "^1.10.1",
-    "prettier": "^2.3.2",
-    "release-it": "^14.12.3",
+    "jest": "^27.5.1",
+    "node-mocks-http": "^1.11.0",
+    "prettier": "^2.5.1",
+    "release-it": "^14.12.4",
     "ts-node-dev": "^1.1.8",
-    "typescript": "4.4.4",
-    "cypress": "^9.0.0"
+    "typescript": "4.5.5",
+    "cypress": "^9.5.0"
   }
 }
