@@ -19,4 +19,4 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
   }
 }
 
-export default jwtVerificationHandler(handler);
+export default jwtVerificationHandler(handler as any);
