@@ -24,6 +24,6 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
   }
 }
 
-export default methodAllowedHandler(jwtVerificationHandler(handler), [
+export default methodAllowedHandler(jwtVerificationHandler(handler as any), [
   HTTPMethod.GET,
 ]);
