@@ -31,6 +31,6 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
 }
 
 export default methodAllowedHandler(
-  jwtVerificationHandler(paginationHandler(handler)),
+  jwtVerificationHandler(paginationHandler(handler as any)),
   [HTTPMethod.GET, HTTPMethod.POST]
 );
